Scene API improvements
